\name{convert2gen}
\alias{convert2gen}

\title{Convert simulated pedigree data to genotypes at markers}

\description{
  Convert data generated by \code{\link{sim.ped}} to a matrix of
  genotypes.
}

\usage{
convert2gen(xodat, map)
}

\arguments{
 \item{xodat}{Allele information with continuous crossover locations, as
 generated by \code{\link{sim.ped}}.}
 \item{map}{A vector of marker locations, in cM.}
}

\details{
  The marker locations (\code{map}) must be sorted and should span a
  length that is less then what was used to generated \code{xodat}.

  If there are just two possible alleles (1 and 2) in \code{xodat}, the
  genotypes are coded as \code{1}, \code{2}, and \code{3}, for genotypes
  11, 12 and 22.  Otherwise, they are converted to binary codes.  For
  example, with possible alleles 1, 2, and 3, an individual with
  genotype a,b will be given the value \eqn{2^(a-1) + 2^(b-1)}.
}  

\value{
  The output is a matrix with \code{length(xodat)} rows and
  \code{length(map)} columns.  Genotype codes are described in Details.
}

\author{Karl W Broman, \email{kbroman@biostat.wisc.edu} }

\examples{
ailped <- genAILped(ngen=2, npairs=2, sibship.size=2)
aildat <- sim.ped(ailped, 100)
map <- seq(0, 100, by=10)
names(map) <- paste("M", 1:length(map), sep="")
ailgen <- convert2gen(aildat, map)
}
                    
\seealso{ \code{\link{sim.ped}}, \code{\link{genAILped}} }

\keyword{datagen}
