\name{ricalcversion}
\alias{ricalcversion}

\title{Installed version of R/ricalc}

\description{
  Print the version number of the currently installed version of R/ricalc.
}

\usage{ricalcversion()}

%\arguments{
%}

\value{
  A character string with the version number of the currently installed
  version of R/ricalc.
}

\author{Karl W Broman, \email{broman@wisc.edu} }

\examples{
  ricalcversion()
}


\keyword{print}
